feat: collapse action and observation messages
