+++ web/src/routes/root/index.css
@@ -4,8 +4,10 @@
 .theme-light {
   --bg-primary: var(--gray-1);
   --bg-secondary: var(--gray-3);
+  --bg-complementary: var(--gray-7);
   --text-primary: var(--gray-8);
   --text-secondary: var(--gray-7);
+  --text-complementary: var(--gray-3);
   --text-3: var(--gray-5);
   --border-color: var(--gray-5);
 }
@@ -13,8 +15,10 @@
 .theme-dark {
   --bg-primary: var(--gray-8);
   --bg-secondary: var(--gray-7);
+  --bg-complementary: var(--gray-3);
   --text-primary: var(--gray-3);
   --text-secondary: var(--gray-4);
+  --text-complementary: var(--gray-8);
   --text-3: var(--gray-6);
   --border-color: var(--gray-6);
 }
